requirements.txt
numpy
pillow
tqdm
torch
torchvision
sklearn